+++ NetworkManager/service/NetworkConnectivity.cpp
@@ -12,14 +12,14 @@
 #include <atomic>
 #include <mutex>
 #include <curl/curl.h>
-
-#include "UtilsLogging.h"
 #include "Module.h"
-#include "Network.h"
 #include "NetworkConnectivity.h"
+#include "NetworkManagerImplementation.h"
 
 namespace WPEFramework {
     namespace Plugin {
+
+    extern NetworkManagerImplementation* _instance;
 
     bool EndpointCache::isEndpointCashFileExist()
     {
@@ -37,11 +37,11 @@ namespace WPEFramework {
                 outputFile << str << '\n';
             }
             outputFile.close();
-            LOGINFO("Connectivity endpoints successfully written to a file");
+            NMLOG_INFO("Connectivity endpoints successfully written to a file");
         }
         else
         {
-            LOGERR("Connectivity endpoints file write error");
+            NMLOG_ERROR("Connectivity endpoints file write error");
         }
     }
 
@@ -60,7 +60,7 @@ namespace WPEFramework {
         }
         else
         {
-            LOGERR("Failed to open connectivity endpoint cache file");
+            NMLOG_ERROR("Failed to open connectivity endpoint cache file");
         }
         return readStrings;
     }
@@ -70,7 +70,7 @@ namespace WPEFramework {
         std::ifstream configFile(configFilePath);
         if (!configFile.is_open()) 
         {
-            LOGERR("Unable to open the configuration file: %s", configFilePath.c_str());
+            NMLOG_ERROR("Unable to open the configuration file: %s", configFilePath.c_str());
             return;
         }
 
@@ -121,15 +121,15 @@ namespace WPEFramework {
 
         if(m_defaultEndpoints.empty())
         {
-            LOGERR("Default endpoints are empty !!");
+            NMLOG_ERROR("Default endpoints are empty !!");
         }
         else
         {
             std::string endpoints_str;
             for (const auto& endpoint : m_defaultEndpoints)
                 endpoints_str.append(endpoint).append(" ");
-            LOGINFO("default endpoints count %d and endpoints:- %s", static_cast<int>(m_defaultEndpoints.size()), endpoints_str.c_str());
-            LOGINFO("default monitor connectivity interval: %d and monitor connectivity auto start : %s", configMonitorInterval, configMonitorConnectivityEnabled?"true":"false");
+            NMLOG_INFO("default endpoints count %d and endpoints:- %s", static_cast<int>(m_defaultEndpoints.size()), endpoints_str.c_str());
+            NMLOG_INFO("default monitor connectivity interval: %d and monitor connectivity auto start : %s", configMonitorInterval, configMonitorConnectivityEnabled?"true":"false");
         }
     }
 
@@ -146,7 +146,7 @@ namespace WPEFramework {
             std::string endpoints_str;
             for (const auto& endpoint : monitorEndpoints)
                 endpoints_str.append(endpoint).append(" ");
-            LOGINFO("updated endpoints count %d and endpoints:- %s", static_cast<int>(monitorEndpoints.size()), endpoints_str.c_str());
+            NMLOG_INFO("updated endpoints count %d and endpoints:- %s", static_cast<int>(monitorEndpoints.size()), endpoints_str.c_str());
             if(monitorEndpoints.size() > 0)
                 return true;
         }
@@ -157,11 +157,11 @@ namespace WPEFramework {
     {
         if (nsm_internetState::FULLY_CONNECTED == getInternetConnectionState(ipversion))
         {
-            LOGINFO("isConnectedToInternet = true");
+            NMLOG_INFO("isConnectedToInternet = true");
             return true;
         }
 
-        LOGWARN("isConnectedToInternet = false");
+        NMLOG_WARNING("isConnectedToInternet = false");
         return false;
     }
 
@@ -200,7 +200,7 @@ namespace WPEFramework {
                 return getCaptivePortal();
             }
             else
-                LOGWARN("No captive portal found !");
+                NMLOG_WARNING("No captive portal found !");
         }
         return std::string("");
     }
@@ -213,7 +213,7 @@ namespace WPEFramework {
             if(!endpoint.empty() && endpoint.size() > 3)
                 monitorEndpoints.push_back(endpoint.c_str());
             else
-                LOGINFO("endpoint not vallied = %s", endpoint.c_str());
+                NMLOG_INFO("endpoint not vallied = %s", endpoint.c_str());
         }
 
         // write the endpoints to a file
@@ -222,7 +222,7 @@ namespace WPEFramework {
         std::string endpoints_str;
         for (const auto& endpoint : monitorEndpoints)
             endpoints_str.append(endpoint).append(" ");
-        LOGINFO("Connectivity monitor endpoints -: %d :- %s", static_cast<int>(monitorEndpoints.size()), endpoints_str.c_str());
+        NMLOG_INFO("Connectivity monitor endpoints -: %d :- %s", static_cast<int>(monitorEndpoints.size()), endpoints_str.c_str());
     }
 
     std::vector<std::string> ConnectivityMonitor::getConnectivityMonitorEndpoints()
@@ -248,7 +248,7 @@ namespace WPEFramework {
     }
     static size_t writeFunction(void* ptr, size_t size, size_t nmemb, std::string* data) {
     #ifdef DBG_CURL_GET_RESPONSE
-        LOG_DBG("%s",(char*)ptr);
+        NMLOG_TRACE("%s",(char*)ptr);
     #endif
         return size * nmemb;
     }
@@ -257,14 +257,14 @@ namespace WPEFramework {
     {
         long deadline = current_time() + timeout_ms, time_now = 0, time_earlier = 0;
         if(endpoints.size() < 1) {
-            LOGERR("endpoints size error ! ");
+            NMLOG_ERROR("endpoints size error ! ");
             return NO_INTERNET;
         }
 
         CURLM *curl_multi_handle = curl_multi_init();
         if (!curl_multi_handle)
         {
-            LOGERR("curl_multi_init returned NULL");
+            NMLOG_ERROR("curl_multi_init returned NULL");
             return NO_INTERNET;
         }
 
@@ -273,12 +273,13 @@ namespace WPEFramework {
         std::vector<int> http_responses;
         struct curl_slist *chunk = NULL;
         chunk = curl_slist_append(chunk, "Cache-Control: no-cache, no-store");
+        chunk = curl_slist_append(chunk, "Connection: close");
         for (const auto& endpoint : endpoints)
         {
             CURL *curl_easy_handle = curl_easy_init();
             if (!curl_easy_handle)
             {
-                LOGERR("endpoint = <%s> curl_easy_init returned NULL", endpoint.c_str());
+                NMLOG_ERROR("endpoint = <%s> curl_easy_init returned NULL", endpoint.c_str());
                 continue;
             }
             curl_easy_setopt(curl_easy_handle, CURLOPT_URL, endpoint.c_str());
@@ -287,7 +288,8 @@ namespace WPEFramework {
             curl_easy_setopt(curl_easy_handle, CURLOPT_HTTPHEADER, chunk);
             curl_easy_setopt(curl_easy_handle, CURLOPT_USERAGENT, "RDKCaptiveCheck/1.0");
             /* set CURLOPT_HTTPGET option insted of CURLOPT_CONNECT_ONLY bcause we need to get the captiveportal URI not just connection only */
-            curl_easy_setopt(curl_easy_handle, CURLOPT_HTTPGET, 1L);
+            /* HTTPGET request added insted of HTTPHEAD request fix for DELIA-61526 */
+            curl_easy_setopt(curl_easy_handle, CURLOPT_HTTPGET, 1L); 
             curl_easy_setopt(curl_easy_handle, CURLOPT_WRITEFUNCTION, writeFunction);
             curl_easy_setopt(curl_easy_handle, CURLOPT_TIMEOUT_MS, deadline - current_time());
             if ((ipversion == CURL_IPRESOLVE_V4) || (ipversion == CURL_IPRESOLVE_V6))
@@ -296,7 +298,7 @@ namespace WPEFramework {
                 curl_easy_setopt(curl_easy_handle, CURLOPT_VERBOSE, 1L);
             if (CURLM_OK != (mc = curl_multi_add_handle(curl_multi_handle, curl_easy_handle)))
             {
-                LOGERR("endpoint = <%s> curl_multi_add_handle returned %d (%s)", endpoint.c_str(), mc, curl_multi_strerror(mc));
+                NMLOG_ERROR("endpoint = <%s> curl_multi_add_handle returned %d (%s)", endpoint.c_str(), mc, curl_multi_strerror(mc));
                 curl_easy_cleanup(curl_easy_handle);
                 continue;
             }
@@ -307,12 +309,12 @@ namespace WPEFramework {
     #if LIBCURL_VERSION_NUM < 0x074200
         int numfds, repeats = 0;
     #endif
-        char *endpoint = nullptr;
+        char *endpoint;
         while (1)
         {
             if (CURLM_OK != (mc = curl_multi_perform(curl_multi_handle, &handles)))
             {
-                LOGERR("curl_multi_perform returned %d (%s)", mc, curl_multi_strerror(mc));
+                NMLOG_ERROR("curl_multi_perform returned %d (%s)", mc, curl_multi_strerror(mc));
                 break;
             }
             for (CURLMsg *msg; NULL != (msg = curl_multi_info_read(curl_multi_handle, &msgs_left)); )
@@ -324,17 +326,17 @@ namespace WPEFramework {
                     curl_easy_getinfo(msg->easy_handle, CURLINFO_PRIVATE, &endpoint);
                     if (curl_easy_getinfo(msg->easy_handle, CURLINFO_RESPONSE_CODE, &response_code) == CURLE_OK)  {
                         if(curlVerboseEnabled())
-                            LOGINFO("endpoint = <%s> http response code <%d>", endpoint, static_cast<int>(response_code));
+                            NMLOG_INFO("endpoint = <%s> response code <%d>",endpoint, static_cast<int>(response_code));
                         if (HttpStatus_302_Found == response_code) {
                             if ( (curl_easy_getinfo(msg->easy_handle, CURLINFO_REDIRECT_URL, &url) == CURLE_OK) && url != nullptr) {
-                                //LOGWARN("captive portal found !!!");
+                                //NMLOG_WARNING("captive portal found !!!");
                                 setCaptivePortal(url);
                             }
                         }
                     }
                 }
                 else
-                    LOGERR("endpoint = <%s> curl error = %d (%s)", endpoint, msg->data.result, curl_easy_strerror(msg->data.result));
+                    NMLOG_TRACE("endpoint = <%s> error = %d (%s)", endpoint, msg->data.result, curl_easy_strerror(msg->data.result));
                 http_responses.push_back(response_code);
             }
             time_earlier = time_now;
@@ -344,7 +346,7 @@ namespace WPEFramework {
     #if LIBCURL_VERSION_NUM < 0x074200
             if (CURLM_OK != (mc = curl_multi_wait(curl_multi_handle, NULL, 0, deadline - time_now, &numfds)))
             {
-                LOGERR("curl_multi_wait returned %d (%s)", mc, curl_multi_strerror(mc));
+                NMLOG_ERROR("curl_multi_wait returned %d (%s)", mc, curl_multi_strerror(mc));
                 break;
             }
             if (numfds == 0)
@@ -358,21 +360,19 @@ namespace WPEFramework {
     #else
             if (CURLM_OK != (mc = curl_multi_poll(curl_multi_handle, NULL, 0, deadline - time_now, NULL)))
             {
-                LOGERR("curl_multi_poll returned %d (%s)", mc, curl_multi_strerror(mc));
+                NMLOG_ERROR("curl_multi_poll returned %d (%s)", mc, curl_multi_strerror(mc));
                 break;
             }
     #endif
         }
-
         if(curlVerboseEnabled()) {
-            LOGINFO("endpoints count = %d response count %d, handles = %d, deadline = %ld, time_now = %ld, time_earlier = %ld",
+            NMLOG_INFO("endpoints count = %d response count %d, handles = %d, deadline = %ld, time_now = %ld, time_earlier = %ld",
                 static_cast<int>(endpoints.size()), static_cast<int>(http_responses.size()), handles, deadline, time_now, time_earlier);
         }
-
         for (const auto& curl_easy_handle : curl_easy_handles)
         {
             curl_easy_getinfo(curl_easy_handle, CURLINFO_PRIVATE, &endpoint);
-            //LOG_DBG("endpoint = <%s> terminating attempt", endpoint);
+            //NMLOG_TRACE("endpoint = <%s> terminating attempt", endpoint);
             curl_multi_remove_handle(curl_multi_handle, curl_easy_handle);
             curl_easy_cleanup(curl_easy_handle);
         }
@@ -418,24 +418,20 @@ namespace WPEFramework {
             {
                 case HttpStatus_204_No_Content:
                     InternetConnectionState = FULLY_CONNECTED;
-                    LOGINFO("Internet State: FULLY_CONNECTED - %.1f%%", (percentage*100));
+                    NMLOG_INFO("Internet State: FULLY_CONNECTED - %.1f%%", (percentage*100));
                 break;
                 case HttpStatus_200_OK:
                     InternetConnectionState = LIMITED_INTERNET;
-                    LOGINFO("Internet State: LIMITED_INTERNET - %.1f%%", (percentage*100));
+                    NMLOG_INFO("Internet State: LIMITED_INTERNET - %.1f%%", (percentage*100));
                 break;
                 case HttpStatus_511_Authentication_Required:
                 case HttpStatus_302_Found:
                     InternetConnectionState = CAPTIVE_PORTAL;
-                    LOGINFO("Internet State: CAPTIVE_PORTAL -%.1f%%", (percentage*100));
+                    NMLOG_INFO("Internet State: CAPTIVE_PORTAL -%.1f%%", (percentage*100));
                 break;
                 default:
                     InternetConnectionState = NO_INTERNET;
-                    if(http_response_code == -1)
-                        LOGERR("Internet State: NO_INTERNET curl error");
-                    else
-                        LOGWARN("Internet State: NO_INTERNET Received http response code: <%d> %.1f%%", static_cast<int>(http_response_code), percentage * 100);
-                    break;
+                    NMLOG_INFO("Internet State: NO_INTERNET Response code: <%d> %.1f%%", static_cast<int>(http_response_code), (percentage*100));
             }
         }
 
@@ -446,7 +442,7 @@ namespace WPEFramework {
     {
         if(!isConnectivityMonitorEndpointSet())
         {
-            LOGINFO("Connectivity monitor endpoints are not set !");
+            NMLOG_INFO("Connectivity monitor endpoints are not set !");
             return false;
         }
 
@@ -456,14 +452,14 @@ namespace WPEFramework {
         {
             isContinuesMonitoringNeeded = true;
             cv_.notify_all();
-            LOGINFO("Connectivity monitor Restarted with %d", timeout.load());
+            NMLOG_INFO("Connectivity monitor Restarted with %d", timeout.load());
             //TODO check still active
         }
         else
         {
             if (thread_.joinable())
             {
-                LOGWARN("Connectivity Monitor joinable Thread is active");
+                NMLOG_WARNING("Connectivity Monitor joinable Thread is active");
                 stopFlag = true;
                 cv_.notify_all();
                 thread_.join();
@@ -473,7 +469,7 @@ namespace WPEFramework {
             stopFlag = false;
             thread_ = std::thread(&ConnectivityMonitor::connectivityMonitorFunction, this);
             threadRunning = true;
-            LOGINFO("Connectivity monitor started with %d", timeout.load());
+            NMLOG_INFO("Connectivity monitor started with %d", timeout.load());
         }
 
         return true;
@@ -483,20 +479,20 @@ namespace WPEFramework {
     {
         if(!isConnectivityMonitorEndpointSet())
         {
-            LOGINFO("Connectivity monitor endpoints are not set !");
+            NMLOG_INFO("Connectivity monitor endpoints are not set !");
             return false;
         }
 
         if (isMonitorThreadRunning())
         {
-            LOGINFO("Connectivity Monitor Thread is active so notify");
+            NMLOG_INFO("Connectivity Monitor Thread is active so notify");
             cv_.notify_all();
         }
         else
         {
             if (thread_.joinable())
             {
-                LOGWARN("Connectivity Monitor joinable Thread is active");
+                NMLOG_WARNING("Connectivity Monitor joinable Thread is active");
                 stopFlag = true;
                 cv_.notify_all();
                 thread_.join();
@@ -506,7 +502,7 @@ namespace WPEFramework {
             timeout.store(timeoutInSeconds >= MONITOR_TIMEOUT_INTERVAL_MIN ? timeoutInSeconds:defaultTimeoutInSec);
             thread_ = std::thread(&ConnectivityMonitor::connectivityMonitorFunction, this);
             threadRunning = true;
-            LOGINFO("Initial Connectivity Monitoring started with %d", timeout.load());
+            NMLOG_INFO("Initial Connectivity Monitoring started with %d", timeout.load());
         }
 
         return true;
@@ -519,31 +515,29 @@ namespace WPEFramework {
 
     bool ConnectivityMonitor::stopInitialConnectivityMonitoring()
     {
-
-        if(isContinuesMonitoringNeeded)
+        if (isMonitorThreadRunning())
         {
-            LOGWARN("Continuous Connectivity Monitor is running");
-            return true;
-        }
-        else
-        {
-            if (!isMonitorThreadRunning())
+            if(isContinuesMonitoringNeeded)
             {
-                LOGWARN("Connectivity monitor not running");
-            }
-
-            stopFlag = true;
-            cv_.notify_all();
-
-            if (thread_.joinable())
-            {
-                thread_.join();
-                threadRunning = false;
-                LOGINFO("Stoping Initial Connectivity Monitor");
+                NMLOG_WARNING("Continuous Connectivity Monitor is running");
+                return true;
             }
             else
-                LOGWARN("thread not joinable !");
+            {
+                stopFlag = true;
+                cv_.notify_all();
+
+                if (thread_.joinable()) {
+                    thread_.join();
+                    threadRunning = false;
+                    NMLOG_INFO("Stoping Initial Connectivity Monitor");
+                }
+                else
+                    NMLOG_WARNING("thread not joinable !");
+            }
         }
+        else
+            NMLOG_WARNING("Continuous Connectivity Monitor not running");
 
         return true;
     }
@@ -552,8 +546,10 @@ namespace WPEFramework {
     {
         if (!isMonitorThreadRunning())
         {
-            LOGWARN("Connectivity monitor not running");
+            NMLOG_WARNING("Connectivity monitor not running");
+            return false;
         }
+
         cv_.notify_all();
         stopFlag = true;
 
@@ -562,10 +558,10 @@ namespace WPEFramework {
             thread_.join();
             isContinuesMonitoringNeeded = false;
             threadRunning = false;
-            LOGINFO("Continuous Connectivity monitor stopped");
+            NMLOG_INFO("Continuous Connectivity monitor stopped");
         }
         else
-            LOGWARN("thread not joinable !");
+            NMLOG_WARNING("thread not joinable !");
         return true;
     }
 
@@ -587,36 +583,50 @@ namespace WPEFramework {
             InternetConnectionState = testConnectivity(getConnectivityMonitorEndpoints(), TEST_CONNECTIVITY_DEFAULT_TIMEOUT_MS, NSM_IPRESOLVE_WHATEVER);
             if(g_internetState.load() != InternetConnectionState)
             {
+                if(_instance != nullptr)
+                {
+                    Exchange::INetworkManager::InternetStatus oldState = static_cast<Exchange::INetworkManager::InternetStatus>(g_internetState.load());
+                    Exchange::INetworkManager::InternetStatus newstate = static_cast<Exchange::INetworkManager::InternetStatus>(InternetConnectionState);
+                    _instance->ReportInternetStatusChangedEvent(oldState , newstate);
+                    NMLOG_TRACE("ReportInternetStatusChangedEvent called");
+                }
+                else
+                    NMLOG_WARNING("ReportInternetStatusChangedEvent callback not set");
+
                 g_internetState.store(InternetConnectionState);
-                Network::notifyInternetStatusChange(g_internetState.load());
             }
 
             if(!isContinuesMonitoringNeeded && (g_internetState.load() == FULLY_CONNECTED))
             {
                 stopFlag = true;
-                LOGINFO("Initial Connectivity Monitoring done Exiting ... Internet state FULLY_CONNECTED");
+                NMLOG_INFO("Initial Connectivity Monitoring done Exiting ... Internet state FULLY_CONNECTED");
                 threadRunning = false;
                 break;
             }
 
             if(stopFlag)
             {
+                NMLOG_WARNING("stopFlag true exiting");
                 threadRunning = false;
                 break;
             }
             //wait for next timout or conditon signal
             std::unique_lock<std::mutex> lock(mutex_);
-            if (cv_.wait_for(lock, std::chrono::seconds(timeout.load())) != std::cv_status::timeout)
+            if (cv_.wait_for(lock, std::chrono::seconds(timeout.load())) == std::cv_status::timeout)
+            {
+                NMLOG_INFO("Connectivity monitor thread timeout");
+            }
+            else
             {
                 if(!stopFlag)
                 {
-                    LOGINFO("Connectivity monitor received a trigger");
+                    NMLOG_INFO("Connectivity monitor received a trigger");
                 }
             }
 
         } while (!stopFlag);
         g_internetState = nsm_internetState::UNKNOWN;
-        LOGWARN("Connectivity monitor exiting");
+        NMLOG_WARNING("Connectivity monitor exiting");
     }
 
     } // namespace Plugin